Replace home screen content into a fragment.
