+++ app/src/main/java/arun/com/chromer/MainActivity.java
@@ -24,36 +24,27 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.Uri;
 import android.os.Bundle;
-import android.speech.RecognizerIntent;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CollapsingToolbarLayout;
 import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v4.content.LocalBroadcastManager;
-import android.support.v4.widget.NestedScrollView;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.CheckBox;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.TextView;
 
 import com.afollestad.materialdialogs.GravityEnum;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.afollestad.materialdialogs.StackingBehavior;
 import com.crashlytics.android.answers.Answers;
 import com.crashlytics.android.answers.CustomEvent;
-import com.crashlytics.android.answers.SearchEvent;
 import com.crashlytics.android.answers.ShareEvent;
 import com.flipboard.bottomsheet.BottomSheetLayout;
 import com.mikepenz.community_material_typeface_library.CommunityMaterial;
-import com.mikepenz.iconics.IconicsDrawable;
 import com.mikepenz.materialdrawer.AccountHeaderBuilder;
 import com.mikepenz.materialdrawer.Drawer;
 import com.mikepenz.materialdrawer.DrawerBuilder;
@@ -64,108 +55,50 @@ import com.mikepenz.materialdrawer.model.SecondaryDrawerItem;
 import java.util.Collections;
 import java.util.List;
 
+import arun.com.chromer.activities.SnackHelper;
 import arun.com.chromer.activities.about.AboutAppActivity;
 import arun.com.chromer.activities.about.changelog.Changelog;
-import arun.com.chromer.activities.history.HistoryActivity;
 import arun.com.chromer.activities.intro.ChromerIntro;
 import arun.com.chromer.activities.intro.WebHeadsIntro;
-import arun.com.chromer.activities.mvp.BaseActivity;
+import arun.com.chromer.activities.main.home.HomeFragment;
 import arun.com.chromer.activities.payments.DonateActivity;
 import arun.com.chromer.activities.settings.Preferences;
 import arun.com.chromer.activities.settings.SettingsGroupActivity;
-import arun.com.chromer.customtabs.CustomTabManager;
 import arun.com.chromer.customtabs.CustomTabs;
-import arun.com.chromer.data.website.WebsiteRepository;
-import arun.com.chromer.data.website.model.WebSite;
-import arun.com.chromer.search.SuggestionItem;
 import arun.com.chromer.shared.Constants;
-import arun.com.chromer.util.RxUtils;
 import arun.com.chromer.util.ServiceUtil;
 import arun.com.chromer.util.Utils;
 import arun.com.chromer.util.cache.FontCache;
-import arun.com.chromer.views.searchview.MaterialSearchView;
-import arun.com.chromer.webheads.WebHeadService;
 import butterknife.BindView;
-import butterknife.OnClick;
+import butterknife.ButterKnife;
 import timber.log.Timber;
 
-import static android.view.View.GONE;
-import static android.view.View.VISIBLE;
 import static arun.com.chromer.shared.Constants.ACTION_CLOSE_ROOT;
-import static arun.com.chromer.shared.Constants.REQUEST_CODE_VOICE;
+import static arun.com.chromer.util.cache.FontCache.MONO;
 
-public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implements Home.View {
+public class MainActivity extends AppCompatActivity implements SnackHelper {
 
     @BindView(R.id.bottomsheet)
     public BottomSheetLayout bottomSheetLayout;
-    @BindView(R.id.material_search_view)
-    public MaterialSearchView materialSearchView;
     @BindView(R.id.toolbar)
     public Toolbar toolbar;
     @BindView(R.id.coordinator_layout)
     public CoordinatorLayout coordinatorLayout;
-    @BindView(R.id.chromer)
-    TextView chromer;
-    @BindView(R.id.incognito_mode)
-    CheckBox incognitoMode;
-    @BindView(R.id.recents_list)
-    RecyclerView recentsList;
-    @BindView(R.id.fab)
-    FloatingActionButton fab;
     @BindView(R.id.root)
     LinearLayout root;
     @BindView(R.id.collapsing_toolbar)
     CollapsingToolbarLayout collapsingToolbar;
     @BindView(R.id.appbar)
     AppBarLayout appbar;
-    @BindView(R.id.nested_scroll_view)
-    NestedScrollView nestedScrollView;
-    @BindView(R.id.recent_missing_text)
-    TextView recentMissingText;
-    @BindView(R.id.history)
-    TextView history;
 
-
-    private CustomTabManager customTabManager;
-    private Drawer drawer;
     private BroadcastReceiver closeReceiver;
-
-    private RecentsAdapter recentsAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         setTheme(R.style.AppTheme_NoActionBar);
         super.onCreate(savedInstanceState);
-
-        setSupportActionBar(toolbar);
-        if (getSupportActionBar() != null) {
-            getSupportActionBar().setHomeButtonEnabled(false);
-            getSupportActionBar().setDisplayHomeAsUpEnabled(false);
-            getSupportActionBar().setDisplayShowHomeEnabled(false);
-            getSupportActionBar().setDisplayShowTitleEnabled(false);
-        }
-        appbar.addOnOffsetChangedListener(new AppBarLayout.OnOffsetChangedListener() {
-            boolean shown = false;
-            int scrollRange = -1;
-
-            @Override
-            public void onOffsetChanged(AppBarLayout appBarLayout, int verticalOffset) {
-                if (scrollRange == -1) {
-                    scrollRange = appBarLayout.getTotalScrollRange();
-                }
-                if (scrollRange + verticalOffset == 0) {
-                    collapsingToolbar.setTitle(getString(R.string.app_name));
-                    chromer.setText(" ");
-                    shown = true;
-                } else if (shown) {
-                    collapsingToolbar.setTitle(" ");
-                    chromer.setText(getString(R.string.app_name));
-                    shown = false;
-                }
-            }
-        });
-        collapsingToolbar.setCollapsedTitleTypeface(FontCache.get(FontCache.MONO, this));
-        chromer.setTypeface(FontCache.get(FontCache.MONO, this));
+        setContentView(R.layout.activity_main);
+        ButterKnife.bind(this);
 
         if (Preferences.get(this).isFirstRun()) {
             startActivity(new Intent(this, ChromerIntro.class));
@@ -173,62 +106,20 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
 
         Changelog.conditionalShow(this);
 
-        setupMaterialSearch();
         setupDrawer();
-        setupCustomTab();
-        setupRecents();
-
         checkAndEducateUser(false);
-
         ServiceUtil.takeCareOfServices(getApplicationContext());
         registerCloseReceiver();
-    }
 
-    private void setupRecents() {
-        recentMissingText.setCompoundDrawablePadding(Utils.dpToPx(8));
-        recentMissingText.setCompoundDrawables(new IconicsDrawable(this)
-                .icon(CommunityMaterial.Icon.cmd_history)
-                .colorRes(R.color.accent)
-                .sizeDp(24), null, null, null);
-        recentsList.setLayoutManager(new GridLayoutManager(this, 4));
-        recentsAdapter = new RecentsAdapter();
-        recentsList.setAdapter(recentsAdapter);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        customTabManager.bindCustomTabsService(this);
-    }
-
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        invalidateState();
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        customTabManager.unbindCustomTabsService(this);
-    }
-
-    @NonNull
-    @Override
-    public Home.Presenter createPresenter() {
-        return new Home.Presenter();
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.fragment_container, new HomeFragment())
+                .commit();
     }
 
     @Override
     protected void onDestroy() {
         LocalBroadcastManager.getInstance(this).unregisterReceiver(closeReceiver);
         super.onDestroy();
-    }
-
-    @Override
-    protected int getLayoutRes() {
-        return R.layout.activity_main;
     }
 
     @Override
@@ -247,20 +138,6 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
         return true;
     }
 
-    private void invalidateState() {
-        incognitoMode.setChecked(Preferences.get(this).incognitoMode());
-        incognitoMode.setCompoundDrawablePadding(Utils.dpToPx(5));
-        incognitoMode.setCompoundDrawables(new IconicsDrawable(this)
-                .icon(CommunityMaterial.Icon.cmd_incognito)
-                .colorRes(R.color.material_dark_color)
-                .sizeDp(24), null, null, null);
-        incognitoMode.setOnCheckedChangeListener((buttonView, isChecked) ->
-                Preferences.get(getApplicationContext()).incognitoMode(isChecked));
-
-        presenter.loadRecents(this);
-    }
-
-
     private void registerCloseReceiver() {
         closeReceiver = new BroadcastReceiver() {
             @Override
@@ -270,55 +147,6 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
             }
         };
         LocalBroadcastManager.getInstance(this).registerReceiver(closeReceiver, new IntentFilter(ACTION_CLOSE_ROOT));
-    }
-
-    private void setupMaterialSearch() {
-        materialSearchView.clearFocus();
-        presenter.registerSearch(materialSearchView.getEditText());
-        materialSearchView.setInteractionListener(new MaterialSearchView.SearchViewInteractionListener() {
-            @Override
-            public void onVoiceIconClick() {
-                Answers.getInstance().logSearch(new SearchEvent().putCustomAttribute("Mode", "Voice"));
-                if (Utils.isVoiceRecognizerPresent(getApplicationContext())) {
-                    startActivityForResult(Utils.getRecognizerIntent(MainActivity.this), REQUEST_CODE_VOICE);
-                } else {
-                    snack(getString(R.string.no_voice_rec_apps));
-                }
-            }
-
-            @Override
-            public void onSearchPerformed(@NonNull String url) {
-                launchCustomTab(url);
-                Answers.getInstance().logSearch(new SearchEvent());
-            }
-
-            @Override
-            public void onHamburgerClick() {
-                if (drawer.isDrawerOpen()) {
-                    drawer.closeDrawer();
-                } else {
-                    drawer.openDrawer();
-                }
-            }
-        });
-
-    }
-
-    @Override
-    public void setSuggestions(@NonNull List<SuggestionItem> suggestions) {
-        materialSearchView.setSuggestions(suggestions);
-    }
-
-    @Override
-    public void setRecents(@NonNull List<WebSite> webSites) {
-        recentsAdapter.setWebSites(webSites);
-        if (webSites.isEmpty()) {
-            recentMissingText.setVisibility(VISIBLE);
-            history.setVisibility(GONE);
-        } else {
-            recentMissingText.setVisibility(GONE);
-            history.setVisibility(VISIBLE);
-        }
     }
 
     @Override
@@ -333,8 +161,13 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
 
 
     private void setupDrawer() {
-        drawer = new DrawerBuilder()
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayShowTitleEnabled(true);
+        collapsingToolbar.setTitle(getString(R.string.app_name));
+        collapsingToolbar.setCollapsedTitleTypeface(FontCache.get(MONO, this));
+        Drawer drawer = new DrawerBuilder()
                 .withActivity(this)
+                .withToolbar(toolbar)
                 .withAccountHeader(new AccountHeaderBuilder()
                         .withActivity(this)
                         .withHeaderBackground(R.drawable.chromer)
@@ -385,7 +218,7 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
                             break;
                         case 3:
                             Utils.openPlayStore(MainActivity.this, getPackageName());
-                            Answers.getInstance().logCustom(new CustomEvent("Send feedback"));
+                            Answers.getInstance().logCustom(new CustomEvent("Rate on Play Store"));
                             break;
                         case 4:
                             startActivity(new Intent(MainActivity.this, ChromerIntro.class));
@@ -431,53 +264,9 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
                     final Intent googleIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.G_COMMUNITY_URL));
                     startActivity(googleIntent);
                 })
-                .onNeutral((dialog, which) -> launchCustomTab(Constants.APP_TESTING_URL))
+                // .onNeutral((dialog, which) -> launchCustomTab(Constants.APP_TESTING_URL))
                 .build()
                 .show();
-    }
-
-    private void launchCustomTab(@Nullable String url) {
-        if (url != null) {
-            if (Preferences.get(this).webHeads()) {
-                if (Utils.isOverlayGranted(this)) {
-                    final Intent webHeadService = new Intent(this, WebHeadService.class);
-                    webHeadService.setData(Uri.parse(url));
-                    startService(webHeadService);
-                } else {
-                    Utils.openDrawOverlaySettings(this);
-                }
-            } else {
-                CustomTabs.from(this)
-                        .forUrl(url)
-                        .withSession(customTabManager.getSession())
-                        .prepare()
-                        .launch();
-            }
-            WebsiteRepository.getInstance(this).getWebsite(url)
-                    .compose(RxUtils.applySchedulers())
-                    .doOnError(Timber::e)
-                    .subscribe();
-        }
-    }
-
-    private void setupCustomTab() {
-        customTabManager = new CustomTabManager();
-        customTabManager.setConnectionCallback(
-                new CustomTabManager.ConnectionCallback() {
-                    @Override
-                    public void onCustomTabsConnected() {
-                        Timber.d("Connected to custom tabs");
-                        try {
-                            customTabManager.mayLaunchUrl(Uri.parse(Constants.GOOGLE_URL), null, null);
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                    }
-
-                    @Override
-                    public void onCustomTabsDisconnected() {
-                    }
-                });
     }
 
     private void checkAndEducateUser(boolean forceShow) {
@@ -504,45 +293,10 @@ public class MainActivity extends BaseActivity<Home.View, Home.Presenter> implem
 
     @Override
     public void onBackPressed() {
-        if (materialSearchView.hasFocus()) {
-            materialSearchView.clearFocus();
-            return;
-        }
-
         if (bottomSheetLayout.isSheetShowing()) {
             bottomSheetLayout.dismissSheet();
             return;
         }
         super.onBackPressed();
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == REQUEST_CODE_VOICE) {
-            switch (resultCode) {
-                case RESULT_OK:
-                    final List<String> resultList = data.getStringArrayListExtra(RecognizerIntent.EXTRA_RESULTS);
-                    if (resultList != null && !resultList.isEmpty()) {
-                        launchCustomTab(Utils.getSearchUrl(resultList.get(0)));
-                    }
-                    break;
-                default:
-            }
-        }
-        super.onActivityResult(requestCode, resultCode, data);
-    }
-
-    @OnClick(R.id.fab)
-    public void onFabClick() {
-        if (materialSearchView.hasFocus() && materialSearchView.getText().length() > 0) {
-            launchCustomTab(materialSearchView.getURL());
-        } else {
-            launchCustomTab(Constants.GOOGLE_URL);
-        }
-    }
-
-    @OnClick(R.id.history)
-    public void onClick() {
-        startActivity(new Intent(this, HistoryActivity.class));
     }
 }
