+++ app/src/main/java/arun/com/chromer/activities/main/home/RecentsAdapter.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package arun.com.chromer;
+package arun.com.chromer.activities.main.home;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -34,6 +34,7 @@ import com.bumptech.glide.Glide;
 import java.util.ArrayList;
 import java.util.List;
 
+import arun.com.chromer.R;
 import arun.com.chromer.activities.BrowserInterceptActivity;
 import arun.com.chromer.data.website.model.WebSite;
 import butterknife.BindView;
@@ -42,10 +43,10 @@ import butterknife.ButterKnife;
 /**
  * Created by arunk on 07-03-2017.
  */
-class RecentsAdapter extends RecyclerView.Adapter<RecentsAdapter.RecentsViewHolder> {
+public class RecentsAdapter extends RecyclerView.Adapter<RecentsAdapter.RecentsViewHolder> {
     private final List<WebSite> webSites = new ArrayList<>();
 
-    RecentsAdapter() {
+    public RecentsAdapter() {
     }
 
     @Override
@@ -81,7 +82,7 @@ class RecentsAdapter extends RecyclerView.Adapter<RecentsAdapter.RecentsViewHold
         return webSites.size();
     }
 
-    void setWebSites(@NonNull List<WebSite> webSites) {
+    public void setWebSites(@NonNull List<WebSite> webSites) {
         this.webSites.clear();
         this.webSites.addAll(webSites);
         notifyDataSetChanged();
