+++ app/src/main/res/layout/activity_main.xml
@@ -28,7 +28,7 @@
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
         android:layout_width="match_parent"
-        android:layout_height="145dp"
+        android:layout_height="wrap_content"
         android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
         <android.support.design.widget.CollapsingToolbarLayout
@@ -39,18 +39,6 @@
             app:expandedTitleMarginEnd="64dp"
             app:expandedTitleMarginStart="48dp"
             app:layout_scrollFlags="scroll|exitUntilCollapsed">
-
-            <TextView
-                android:id="@+id/chromer"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:gravity="center"
-                android:paddingBottom="24dp"
-                android:paddingTop="24dp"
-                android:text="@string/app_name"
-                android:textColor="@android:color/white"
-                android:textSize="34sp" />
 
             <android.support.v7.widget.Toolbar
                 android:id="@+id/toolbar"
@@ -63,100 +51,11 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <android.support.v4.widget.NestedScrollView
-        android:id="@+id/nested_scroll_view"
+    <FrameLayout
+        android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-                <arun.com.chromer.views.searchview.MaterialSearchView
-                    android:id="@+id/material_search_view"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="top" />
-
-                <CheckBox
-                    android:id="@+id/incognito_mode"
-                    android:layout_width="wrap_content"
-                    android:layout_height="36dp"
-                    android:layout_gravity="center"
-                    android:clickable="true"
-                    android:gravity="center"
-                    android:text="@string/incognito_mode"
-                    android:textColor="@color/material_dark_color"
-                    android:textSize="16sp"
-                    android:visibility="gone" />
-
-                <android.support.v7.widget.CardView
-                    android:id="@+id/recents_card"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginLeft="10dp"
-                    android:layout_marginRight="10dp"
-                    app:cardPreventCornerOverlap="true"
-                    app:cardUseCompatPadding="true">
-
-                    <LinearLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:orientation="vertical"
-                        android:padding="10dp">
-
-                        <TextView
-                            android:id="@+id/recent_missing_text"
-                            android:layout_width="wrap_content"
-                            android:layout_height="36dp"
-                            android:layout_gravity="center"
-                            android:clickable="true"
-                            android:gravity="center"
-                            android:text="@string/recents_intro"
-                            android:textColor="@color/material_dark_color"
-                            android:textSize="16sp"
-                            android:visibility="visible" />
-
-                        <android.support.v7.widget.RecyclerView
-                            android:id="@+id/recents_list"
-                            android:layout_width="match_parent"
-                            android:layout_height="wrap_content"
-                            android:clipToPadding="false" />
-
-                        <TextView
-                            android:id="@+id/history"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_gravity="center|end"
-                            android:background="?attr/selectableItemBackground"
-                            android:clickable="true"
-                            android:gravity="center|end"
-                            android:padding="8dp"
-                            android:text="@string/history_link"
-                            android:textAllCaps="true"
-                            android:textAppearance="@style/TextAppearance.AppCompat.Widget.Button"
-                            android:textColor="@color/accent"
-                            android:textSize="14sp" />
-                    </LinearLayout>
-                </android.support.v7.widget.CardView>
-            </LinearLayout>
-        </FrameLayout>
-
-    </android.support.v4.widget.NestedScrollView>
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@drawable/ic_launch_white_24dp" />
+        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
     <include layout="@layout/widget_bottom_sheet_layout" />
 </android.support.design.widget.CoordinatorLayout>
